requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.9   # for file uploads
openai==2.2.0
